@@ -15,8 +15,8 @@
 		<angular.project.location>.</angular.project.location>
 		<angular.project.nodeinstallation>../node_installation</angular.project.nodeinstallation>
 
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
    <build>
@@ -25,7 +25,7 @@
 			<plugin>
 				<groupId>com.github.eirslett</groupId>
 				<artifactId>frontend-maven-plugin</artifactId>
-				<version>1.12.0</version>
+				<version>${frontend-maven-plugin.version}</version>
 				<configuration>
 					<workingDirectory>${angular.project.location}</workingDirectory>
 					<installDirectory>${angular.project.nodeinstallation}</installDirectory>
@@ -172,12 +172,12 @@
 					<plugin>
 						<groupId>com.spotify</groupId>
 						<artifactId>docker-maven-plugin</artifactId>
-						<version>1.2.2</version>
+						<version>${docker-maven-plugin.version}</version>
 						<dependencies>
 							<dependency>
 								<groupId>com.google.guava</groupId>
 								<artifactId>guava</artifactId>
-								<version>28.2-jre</version>
+								<version>${guava.version}</version>
 							</dependency>
 						</dependencies>
 						<executions>
