@@ -6,19 +6,19 @@ package org.georchestra.signalement.core.entity.reporting;
 import java.util.Date;
 import java.util.UUID;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-import javax.persistence.JoinColumn;
-import javax.persistence.Lob;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.InheritanceType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 import org.georchestra.signalement.core.common.LongId;
 import org.georchestra.signalement.core.dto.GeographicType;
@@ -86,6 +86,10 @@ public abstract class AbstractReportingEntity implements LongId {
 
 	public AbstractReportingEntity(GeographicType geographicType) {
 		this.geographicType = geographicType;
+	}
+
+	public AbstractReportingEntity() {
+
 	}
 
 	public abstract void setGeometry(Geometry geometry);
