@@ -3,9 +3,9 @@
  */
 package org.georchestra.signalement.core.entity.reporting;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
 
 import org.georchestra.signalement.core.dto.GeographicType;
 import org.locationtech.jts.geom.Geometry;
