@@ -3,9 +3,17 @@
  */
 package org.georchestra.signalement.core.entity.acl;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Data;
 
-import javax.persistence.*;
 
 /**
  * Représente l'association d'un utilisateur avec un rôle pour un contexte
